+++ Vehicles/src/com/endava/commands/creation/CreateTrainParkingCommand.java
@@ -18,7 +18,7 @@ public class CreateTrainParkingCommand implements Command {
   private final VehiclesRepository repository;
   private String name;
   private int capacity;
-  private BigDecimal hourlyRate;
+  private double hourlyRate;
 
   public CreateTrainParkingCommand(VehiclesFactory factory, VehiclesRepository repository) {
     this.factory = factory;
@@ -44,7 +44,7 @@ public class CreateTrainParkingCommand implements Command {
     try {
       name = parameters.get(0);
       capacity = Integer.parseInt(parameters.get(1));
-      hourlyRate = BigDecimal.valueOf(Double.parseDouble(parameters.get(2)));
+      hourlyRate = Double.parseDouble(parameters.get(2));
     } catch (Exception e) {
       throw new IllegalArgumentException("Failed to parse CreateTrainParking command parameters.");
     }
