+++ Vehicles/src/com/endava/models/parkings/CarParkingImpl.java
@@ -1,13 +1,11 @@
 package com.endava.models.parkings;
 
-import com.endava.models.parkings.contracts.AirplaneParking;
 import com.endava.models.parkings.contracts.CarParking;
 import com.endava.models.parkings.enums.ParkingType;
-import java.math.BigDecimal;
 
 public class CarParkingImpl extends ParkingBase implements CarParking {
 
-  public CarParkingImpl(String name, int capacity, BigDecimal hourlyRate) {
+  public CarParkingImpl(String name, int capacity, double hourlyRate) {
     super(name, capacity, hourlyRate, ParkingType.CAR);
   }
 }
